+++ src/path.h
@@ -6,6 +6,7 @@
 char *path_prj_dir(char *base, const tman_arg_t * args);
 char *path_brd_dir(char *base, const tman_arg_t * args);
 char *path_task_dir(char *base, const tman_arg_t * args);
+char *path_prj_board(const char *base, const tman_arg_t * args);
 
 char *path_prj_unit(char *base, const tman_arg_t * args);
 char *path_brd_unit(char *base, const tman_arg_t * args);